// packages/client/src/defs.d.ts
export declare type PrismaAppSyncOptionsType = {
    connectionString?: string;
    sanitize?: boolean;
    debug?: boolean;
    defaultPagination?: number | false;
    maxDepth?: number;
};
export declare type Options = Required<PrismaAppSyncOptionsType> & {
    modelsMapping: any;
};
export declare type InjectedConfig = {
    modelsMapping?: {
        [modelVariant: string]: string;
    };
    operations?: string;
};
export declare type Action = typeof Actions[keyof typeof Actions] | string;
export declare type ActionsAlias = typeof ActionsAliases[keyof typeof ActionsAliases] | 'custom' | null;
export declare type ActionsAliasStr = keyof typeof ActionsAliases;
export declare type Context = {
    action: Action;
    alias: ActionsAlias;
    model: string | null;
};
// ...
export declare type QueryParams = {
    type: GraphQLType;
    operation: string;
    context: Context;
    fields: string[];
    paths: string[];
    args: any;
    prismaArgs: PrismaArgs;
    authorization: Authorization;
    identity: Identity;
    headers: any;
};
export declare type Authorization = typeof Authorizations[keyof typeof Authorizations] | null;
// ...
export declare type Reason = string | ((context: Context) => string);
export declare type Shield = {
    [matcher: string]: boolean | {
        rule: boolean | any;
        reason?: Reason;
    };
};
// ...
export declare type ShieldAuthorization = {
    canAccess: boolean;
    reason: Reason;
    prismaFilter: any;
    matcher: string;
    globPattern: string;
};
export declare type ResolveParams<Operations extends string, CustomResolvers extends string> = {
    event: AppsyncEvent;
    resolvers?: {
        [resolver in CustomResolvers]: ((props: QueryParamsCustom) => Promise<any>) | boolean;
    };
    shield?: (props: QueryParams) => Shield;
    hooks?: Hooks<Operations, CustomResolvers>;
};
export { PrismaClient };
export declare type PrismaArgs = {
    where?: any;
    data?: any;
    orderBy?: any;
    skip?: number;
    take?: number;
    skipDuplicates?: boolean;
    select?: any;
};
export declare type AppsyncEvent = {
    arguments: any;
    source: any;
    identity: Identity;
    request: any;
    info: {
        fieldName: string;
        parentTypeName: string;
        variables: any;
        selectionSetList: string[];
        selectionSetGraphQL: string;
    };
    prev: {
        result: any;
    };
    stash: any;
};
export declare type GraphQLType = 'Query' | 'Mutation' | 'Subscription';
export declare type API_KEY = null | {
    [key: string]: any;
};
export declare type AWS_LAMBDA = {
    resolverContext: any;
    [key: string]: any;
};
export declare type AWS_IAM = {
    accountId: string;
    cognitoIdentityPoolId: string;
    cognitoIdentityId: string;
    sourceIp: string[];
    username: string;
    userArn: string;
    cognitoIdentityAuthType: string;
    cognitoIdentityAuthProvider: string;
    [key: string]: any;
};
export declare type AMAZON_COGNITO_USER_POOLS = {
    sub: string;
    issuer: string;
    username: string;
    claims: any;
    sourceIp: string[];
    defaultAuthStrategy: string;
    groups: string[];
    [key: string]: any;
};
export declare type OPENID_CONNECT = {
    claims: {
        sub: string;
        aud: string;
        azp: string;
        iss: string;
        exp: number;
        iat: number;
        gty: string;
    };
    sourceIp: string[];
    issuer: string;
    sub: string;
    [key: string]: any;
};
export declare type Identity = API_KEY | AWS_LAMBDA | AWS_IAM | AMAZON_COGNITO_USER_POOLS | OPENID_CONNECT;
export declare const ReservedPrismaKeys: string[];
export declare enum Actions {
    get = "get",
    list = "list",
    count = "count",
    createMany = "createMany",
    updateMany = "updateMany",
    deleteMany = "deleteMany",
    create = "create",
    update = "update",
    upsert = "upsert",
    delete = "delete",
    onCreatedMany = "onCreatedMany",
    onUpdatedMany = "onUpdatedMany",
    onDeletedMany = "onDeletedMany",
    onMutatedMany = "onMutatedMany",
    onCreated = "onCreated",
    onUpdated = "onUpdated",
    onUpserted = "onUpserted",
    onDeleted = "onDeleted",
    onMutated = "onMutated"
}
export declare enum ActionsAliases {
    access = "access",
    batchAccess = "batchAccess",
    create = "create",
    batchCreate = "batchCreate",
    delete = "delete",
    batchDelete = "batchDelete",
    modify = "modify",
    batchModify = "batchModify",
    subscribe = "subscribe",
    batchSubscribe = "batchSubscribe"
}
// ...
export declare const ActionsList: string[];
export declare const BatchActionsList: string[];
export declare const DebugTestingKey = "__prismaAppsync";
// ...
export declare enum Authorizations {
    API_KEY = "API_KEY",
    AWS_IAM = "AWS_IAM",
    AMAZON_COGNITO_USER_POOLS = "AMAZON_COGNITO_USER_POOLS",
    AWS_LAMBDA = "AWS_LAMBDA",
    OPENID_CONNECT = "OPENID_CONNECT"
}